requirements.txt
discord.py
flask
requests
discord-py-interactions